Removed function tests strdt03 and strlang03 to align with RDF 1.1
The SPARQL 1.1 test suite expects "abc"^^xsd:string and
"DEF"^^xsd:string to produce an error when passed to STRDT and
STRLANG, leading to ?str1 being unbound in these tests. This is
because STRDT and STRLANG expect their first argument to a "simple
literal". However, in RDF 1.1 "simple literals" are merely syntactic
sugar for xsd:string datatyped literals. Therefore, both of these
literals should be acceptable to STRDT and STRLANG, yielding values
bound to ?str1.