Switches turtle list positive syntax tests to evaluation tests
Tests turtle-eval-lists-01 to turtle-eval-lists-05 were previously
positive syntax tests. These tests are now evaluation tests, so for
each Turtle lists test there is a corresponding NTriples .nt file with
the expected output. This change still checks for Turtle collections
syntax compliance as before, and it now also checks that blank nodes
are generated correctly when the `( )` shorthand is used for
collections as a subject or as an object.